requirements.txt
python-dotenv==1.0.0
robonomics-interface==1.6.1
websocket-client==1.6.1
termcolor==2.4.0
flask-classful==0.16.0